reorganized server and delegate, target responsibilities

+++ hopserver/dp_proxy.go
@@ -13,8 +13,11 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// Agproxy holds state used by server to proxy delegate requests to principals
-type Agproxy struct {
+/* DP Proxy: Server acts as a proxy between the Delegate client
+ * and Principal client. */
+
+// dpproxy holds state used by server to proxy delegate requests to principals
+type dpproxy struct {
 	address    string
 	principals map[int32]*hopSession // TODO(baumanl): pointer to hopSession may not work how I want...
 	running    bool
@@ -22,27 +25,24 @@ type Agproxy struct {
 	proxyLock  sync.Mutex
 }
 
-// Running returns whether agproxy is currently accepting connections
-func (p *Agproxy) Running() bool {
+// start starts agproxy server
+func (p *dpproxy) start() error {
 	p.proxyLock.Lock()
 	defer p.proxyLock.Unlock()
-	return p.running
-}
 
-// Start starts agproxy server
-func (p *Agproxy) Start() error {
-	p.proxyLock.Lock()
-	defer p.proxyLock.Unlock()
+	if p.running {
+		return fmt.Errorf("agproxy: start called when already running")
+	}
 
 	fileInfo, err := os.Stat(p.address)
 	if err == nil { // file exists
 		// IsDir is short for fileInfo.Mode().IsDir()
 		if fileInfo.IsDir() {
-			return fmt.Errorf("Agproxy: unable to start agproxy at address %s: is a directory", p.address)
+			return fmt.Errorf("agproxy: unable to start agproxy at address %s: is a directory", p.address)
 		}
 		//make sure the socket does not already exist.
 		if err := os.RemoveAll(p.address); err != nil {
-			return fmt.Errorf("Agproxy: error removing %s: %s", p.address, err)
+			return fmt.Errorf("agproxy: error removing %s: %s", p.address, err)
 		}
 	}
 
@@ -50,7 +50,7 @@ func (p *Agproxy) Start() error {
 	sockconfig := &net.ListenConfig{Control: setListenerOptions}
 	authgrantServer, err := sockconfig.Listen(context.Background(), "unix", p.address)
 	if err != nil {
-		return fmt.Errorf("Agproxy: unix socket listen error: %s", err)
+		return fmt.Errorf("agproxy: unix socket listen error: %s", err)
 	}
 
 	p.listener = authgrantServer
@@ -61,7 +61,7 @@ func (p *Agproxy) Start() error {
 }
 
 // serve accepts connections and starts proxying
-func (p *Agproxy) serve() {
+func (p *dpproxy) serve() {
 	logrus.Info("Agproxy: listening on unix socket: ", p.listener.Addr().String())
 	for {
 		c, err := p.listener.Accept()
@@ -74,7 +74,7 @@ func (p *Agproxy) serve() {
 }
 
 // checks that the connecting process is a hop session descendent and then proxies
-func (p *Agproxy) checkAndProxy(c net.Conn) {
+func (p *dpproxy) checkAndProxy(c net.Conn) {
 	// Verify that the client is a legit descendent and get principal sess
 	principalSess, e := p.checkCredentials(c)
 	if e != nil {
@@ -86,7 +86,7 @@ func (p *Agproxy) checkAndProxy(c net.Conn) {
 
 // verifies that client is a descendent of a process started by the principal
 // and returns its corresponding principal session
-func (p *Agproxy) checkCredentials(c net.Conn) (*hopSession, error) {
+func (p *dpproxy) checkCredentials(c net.Conn) (*hopSession, error) {
 	// cPID is PID of client process that connected to socket
 	cPID, err := readCreds(c)
 	if err != nil {
@@ -127,7 +127,7 @@ func checkDescendents(tree *pstree.Tree, proc pstree.Process, cPID int) bool {
 
 // proxyAuthGrantRequest is used by Server to forward INTENT_REQUESTS from a Client -> Principal and responses from Principal -> Client
 // Checks hop client process is a descendent of the hop server and conducts authgrant request with the appropriate principal
-func (p *Agproxy) proxyAuthGrantRequest(principalSess *hopSession, c net.Conn) {
+func (p *dpproxy) proxyAuthGrantRequest(principalSess *hopSession, c net.Conn) {
 	defer c.Close()
 	if principalSess.transportConn.IsClosed() {
 		logrus.Error("Agproxy: connection with principal is closed or closing")
